pyproject.toml
[project]
name = "filmreviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
]
